+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 ## [Unreleased]
 
 
+## [0.24.4] - 2019-03-08
+### Changed
+- [EVM] Accept []byte nonce rather than enforcing the use of txs.Tx.TxHash()
+- [Crypto] Expose SequenceNonce helper to allow library users to use sequence-number based addresses for newly created contracts
+
+
 ## [0.24.3] - 2019-03-06
 ### Fixed
 - [State] Avoid stack traces which may be code-path-dependent or non-deterministic from being pushed to TxExecutions and so to merkle state where they can lead to breaking consensus
@@ -408,7 +414,8 @@ This release marks the start of Eris-DB as the full permissioned blockchain node
   - [Blockchain] Fix getBlocks to respect block height cap.
 
 
-[Unreleased]: https://github.com/hyperledger/burrow/compare/v0.24.3...HEAD
+[Unreleased]: https://github.com/hyperledger/burrow/compare/v0.24.4...HEAD
+[0.24.4]: https://github.com/hyperledger/burrow/compare/v0.24.3...v0.24.4
 [0.24.3]: https://github.com/hyperledger/burrow/compare/v0.24.2...v0.24.3
 [0.24.2]: https://github.com/hyperledger/burrow/compare/v0.24.1...v0.24.2
 [0.24.1]: https://github.com/hyperledger/burrow/compare/v0.24.0...v0.24.1